requirements.txt
asgiref==3.5.1
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.1
distlib==0.3.6
distro==1.7.0
Django==4.0.5
django-admin==2.0.1
django-environ==0.9.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==22.1
django-six==1.0.4
djangorestframework==3.13.1
djangorestframework-jsonapi==6.0.0
drf-spectacular==0.23.1
exceptiongroup==1.1.0
filelock==3.9.0
flake8==6.0.0
future==0.18.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.1.0
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.8.0
Jinja2==3.0.3
jsonschema==3.2.0
keyring==23.5.0
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==23.0
Pillow==9.0.1
platformdirs==3.1.0
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyrsistent==0.18.1
pytest==7.2.2
pytz==2022.1
PyYAML==5.4.1
requests==2.28.1
screen==1.0.1
SecretStorage==3.3.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.13
validators==0.20.0
virtualenv==20.20.0
xlwt==1.3.0
zipp==3.11.0
